Introduce TestRunner and TimingProbe interfaces
