+++ src/main/java/com/github/protocolfuzzing/protocolstatefuzzer/statefuzzer/testrunner/timingprobe/TimingProbeBuilder.java
@@ -1,13 +1,11 @@
 package com.github.protocolfuzzing.protocolstatefuzzer.statefuzzer.testrunner.timingprobe;
 
-import com.github.protocolfuzzing.protocolstatefuzzer.components.sul.mapper.abstractsymbols.MapperInput;
-import com.github.protocolfuzzing.protocolstatefuzzer.components.sul.mapper.abstractsymbols.MapperOutput;
 import com.github.protocolfuzzing.protocolstatefuzzer.statefuzzer.testrunner.timingprobe.config.TimingProbeEnabler;
 
 /**
  * Builder Interface for the TimingProbe.
  */
-public interface TimingProbeBuilder<S, I extends MapperInput<S, I, O, P>, O extends MapperOutput<O, P>, P> {
+public interface TimingProbeBuilder {
 
     /**
      * Builds a new TimingProbe instance.
@@ -15,5 +13,5 @@ public interface TimingProbeBuilder<S, I extends MapperInput<S, I, O, P>, O exte
      * @param timingProbeEnabler  the configuration that enables the timing probe testing
      * @return                    a new TimingProbe instance
      */
-    TimingProbe<S, I, O, P> build(TimingProbeEnabler timingProbeEnabler);
+    TimingProbe build(TimingProbeEnabler timingProbeEnabler);
 }
